Use Pandas instead of Numpy to read and store datasets. Implement GE_GENERATE.
In #145, @cgarcia-UCO had some nice ideas for new auto-generated grammar content. We discussed it a bit more and we had ideas for users writing executable code inside the grammar. I implemented that. And it turned out to require another change, storing our datasets as Pandas DataFrames, because it requires a datatype per column. That has knock-on effects, including in our bnf files. I've implemented all that. I'm not sure if all this would be acceptable to everyone... so, just making a PR for discussion, or at least to hold this code for posterity.